# openapi.yaml
openapi: 3.1.0
info:
  title: 山海灵感便利店领域 API
  version: 2.0.0
  description: App 用户接口与管理端事务动作。字段统一 snake_case，金额统一十进制字符串。
servers:
  - url: http://localhost:3000/api/v1
...
paths:
...
  /products:
    get:
      tags: [Catalog]
      summary: 商品分页
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: category, in: query, schema: { type: string } }
        - { name: keyword, in: query, schema: { type: string, maxLength: 64 } }
        - { name: tag, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string, enum: [newest, sales, price_asc, price_desc] } }
      responses: { "200": { $ref: "#/components/responses/List" } }
...
  /scan/barcodes/{barcode}:
    get:
      parameters: [{ name: barcode, in: path, required: true, schema: { type: string } }]
      summary: 本地与缓存条码查询
      responses: { "200": { $ref: "#/components/responses/Success" }, "404": { $ref: "#/components/responses/Error" } }
  /scan/products:
    post:
      security: [{ UserBearer: [] }]
      summary: 提交待审扫码商品
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [barcode, name, price], properties: { barcode: { type: string }, name: { type: string }, price: { type: string }, category_id: { $ref: "#/components/schemas/Id" }, cover_image: { type: string, format: uri } } } } } }
      responses: { "201": { $ref: "#/components/responses/Success" } }
  /scan/products/mine:
    get:
      security: [{ UserBearer: [] }]
      summary: 我的扫码提交
      responses: { "200": { $ref: "#/components/responses/List" } }
...
components:
  securitySchemes:
    UserBearer: { type: http, scheme: bearer, description: 登录返回的 App 会话 Token }
    AdminBearer: { type: http, scheme: bearer, description: 环境变量 API_TOKEN }
...
  schemas:
    Id: { oneOf: [{ type: string, pattern: '^\d+$' }, { type: integer, minimum: 1 }] }
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties: { code: { type: string }, message: { type: string }, details: {} }
        requestId: { type: string }
  responses:
    Success:
      description: 成功
      content: { application/json: { schema: { type: object, properties: { data: {} } } } }
    List:
      description: 分页成功
      content:
        application/json:
          schema:
            type: object
            properties:
              data: { type: array, items: {} }
              meta: { type: object, properties: { page: { type: integer }, pageSize: { type: integer }, total: { type: integer } } }
    Error:
      description: 请求失败
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }